pyproject.toml
[project]
name = "webbuddy"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladimir Lamkin",email = "lastloony@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8,<6.0.0",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv (>=1.0.0)",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.1",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
